src/requirements.txt
google-api-core
google-auth
google-cloud-bigquery
google-cloud-discoveryengine
fastapi
pydantic
pyyaml
uvicorn
